pyproject.toml
[project]
name = "julienpillaud"
# use dynamic version to avoid sync issues with uv.lock after 'release-please'
dynamic = ["version"]
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]==0.136.0",
    "logfire[fastapi]==4.32.1",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt==2.12.1",
    "pymongo==4.16.0",
]

[dependency-groups]
dev = [
    "mypy==1.20.1",
    "pre-commit==4.5.1",
    "pytest==9.0.3",
    "ruff==0.15.11",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
output-format = "concise"

[tool.ruff.lint]
select = [
    "E", "W", # pycodestyle
    "F", # Pyflakes
    "UP", # pyupgrade
    "B", # flake8-bugbear
    "SIM", # flake8-simplify
    "I", # isort
    "PL", # Pylint
    "RUF",
]

[tool.mypy]
strict = true
disable_error_code = [
    # Decorators on top of @property are not supported,
    # but pydantic use it
    "prop-decorator"
]
plugins = ['pydantic.mypy']

[tool.pytest.ini_options]
testpaths = ["tests"]
log_cli = true
log_cli_level = "info"
log_cli_format = "%(asctime)s [%(levelname)8s] %(name)30s - %(message)s"
